requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
docopt==0.6.2
flake8==4.0.1
idna==3.3
mccabe==0.6.1
notifications-python-client==6.3.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.0
toml==0.10.2
urllib3==1.26.9
